+++ CHANGES.rst
@@ -9,6 +9,9 @@ There's a frood who really knows where his towel is.
 .. Warning::
     Upgrades are supported only from release 1.0b3.
 
+- Restore default binding on Link content type at uninstall time.
+  [hvelarde]
+
 - Add ``Current`` as default value for available_genres and ``General`` as
   default value for available_sections.
   [hvelarde]